requirements.txt
transformers>=4.41.2,<=4.42.4
datasets>=2.16.0,<=2.20.0
accelerate>=0.30.1,<=0.33.0
peft>=0.11.1,<=0.12.0
trl>=0.8.6,<=0.9.6
torch>=2.2.0,<=2.4.0
vllm>=0.4.3,<0.5.2
bitsandbytes>=0.39.0
autoawq
gradio>=4.0.0
pandas>=2.0.0
scipy
einops
sentencepiece
tiktoken
protobuf
uvicorn
pydantic
fastapi
sse-starlette
matplotlib>=3.7.0
fire
packaging
pyyaml
numpy<2.0.0
optimum>=1.16.0
